Move dataValidator
