+++ .eslintrc.js
@@ -24,6 +24,6 @@ module.exports = {
     'linebreak-style': process.platform === 'win32' ? 'off' : ['error', 'unix'],
     'react-hooks/rules-of-hooks': 'error',
     'react-hooks/exhaustive-deps': 'warn',
-    'import/no-extraneous-dependencies': ['off', {'devDependencies': ['/.storybook/**', '**/stories.jsx', '/src/app/helpers/tests/**', '/src/app/dataValidator/**']}],
+    'import/no-extraneous-dependencies': ['off', {'devDependencies': ['/.storybook/**', '**/stories.jsx', '/src/app/helpers/tests/**', '/dataValidator/**']}],
   },
 };
