+++ dataValidator/helpers/dataLoader/readScenario.js
