+++ dataValidator/helpers/validators/validateNode.js
