+++ dataValidator/helpers/validators/validateType.js
