+++ dataValidator/utilities/countScenarios.js
